requirements.txt
asgiref==3.7.2
Django==4.2.7
sqlparse==0.4.4
tzdata==2023.3
django-colorfield==0.10.1
django-debug-toolbar==4.2
python-dotenv==1.0.0
requests
django-markdownify